Implemented buy Interface
Had to make many corrections due to double inaccuracy, as well as
implemting additional safety checks. For example whether or not a
sellerID is actually held by a user with a seller role.
Next to do would probably be altering all database objects whenever a
user changes role. For a user with a buyer role can't have a sellerID in
a product object. This would need to be corrected automatically.